Update cppo to support Netatalk 3.x and newer
This update adds the `-ea` switch which allows one to create files compatible with Netatalk 3.0 and newer's 'ea' AppleDouble system.

This writes all metadata to extended attributes on the file and the resource fork to a MacOS X style '._' file.

This update requires Python 3.6 due to the use of f-strings.