Updates for Debian Trixie
Update cppo to deal with utcfromtimestamp() deprecation

Python 3.12 is complaining about this.

Install ulockmgr_server if needed

Debian Trixie no longer ships with the ulockmgr_server that was included with FUSE. This is needed in order for Netatalk to properly lock files on shares using ciopfs.

Use ip tuntap to setup tash0

tunctl is no longer available due to deprecation of uml-utilities.